Move CPI functions to financials module
Consolidate these funcs with investment cost calcs in a single location.

Raise depreciation warning for users who are explicitly calling from
the old location.
